完善了联盟功能，修改了一些bug

+++ hall/models/response/union.go
@@ -24,24 +24,24 @@ type JoinUnionResp struct {
 	UpdateUserData map[string]any `json:"updateUserData"`
 }
 type UnionMember struct {
-	Uid                       string `json:"uid"`
-	Nickname                  string `json:"nickname"`
-	Avatar                    string `json:"avatar"`
-	FrontendId                string `json:"frontendId"`
-	RoomId                    string `json:"roomId"`
-	SpreaderID                string `json:"spreaderID" bson:"spreaderID"`                               //推广员ID
-	ProhibitGame              bool   `json:"prohibitGame" bson:"prohibitGame"`                           // 禁止游戏
-	RebateRate                int    `json:"rebateRate" bson:"rebateRate"`                               // 返利比例
-	YesterdayDraw             int    `bson:"yesterdayDraw" json:"yesterdayDraw"`                         // 昨日总局数
-	YesterdayBigWinDraw       int    `bson:"yesterdayBigWinDraw" json:"yesterdayBigWinDraw"`             // 昨日大赢家局数
-	YesterdayRebate           int    `bson:"yesterdayRebate" json:"yesterdayRebate"`                     // 昨日返利
-	TodayRebate               int    `bson:"todayRebate" json:"todayRebate"`                             // 今日返利
-	TotalDraw                 int    `bson:"totalDraw" json:"totalDraw"`                                 // 总局数
-	MemberYesterdayDraw       int    `bson:"memberYesterdayDraw" json:"memberYesterdayDraw"`             // 成员昨日总局数
-	MemberYesterdayBigWinDraw int    `bson:"memberYesterdayBigWinDraw" json:"memberYesterdayBigWinDraw"` // 成员昨日大赢家局数
-	YesterdayProvideRebate    int    `bson:"yesterdayProvideRebate" json:"yesterdayProvideRebate"`       // 昨日贡献返利
-	Score                     int    `json:"score"`
-	SafeScore                 int    `json:"safeScore"`
+	Uid                       string  `json:"uid"`
+	Nickname                  string  `json:"nickname"`
+	Avatar                    string  `json:"avatar"`
+	FrontendId                string  `json:"frontendId"`
+	RoomId                    string  `json:"roomId"`
+	SpreaderID                string  `json:"spreaderID" bson:"spreaderID"`                               //推广员ID
+	ProhibitGame              bool    `json:"prohibitGame" bson:"prohibitGame"`                           // 禁止游戏
+	RebateRate                float64 `json:"rebateRate" bson:"rebateRate"`                               // 返利比例
+	YesterdayDraw             int     `bson:"yesterdayDraw" json:"yesterdayDraw"`                         // 昨日总局数
+	YesterdayBigWinDraw       int     `bson:"yesterdayBigWinDraw" json:"yesterdayBigWinDraw"`             // 昨日大赢家局数
+	YesterdayRebate           int     `bson:"yesterdayRebate" json:"yesterdayRebate"`                     // 昨日返利
+	TodayRebate               int     `bson:"todayRebate" json:"todayRebate"`                             // 今日返利
+	TotalDraw                 int     `bson:"totalDraw" json:"totalDraw"`                                 // 总局数
+	MemberYesterdayDraw       int     `bson:"memberYesterdayDraw" json:"memberYesterdayDraw"`             // 成员昨日总局数
+	MemberYesterdayBigWinDraw int     `bson:"memberYesterdayBigWinDraw" json:"memberYesterdayBigWinDraw"` // 成员昨日大赢家局数
+	YesterdayProvideRebate    int     `bson:"yesterdayProvideRebate" json:"yesterdayProvideRebate"`       // 昨日贡献返利
+	Score                     int     `json:"score"`
+	SafeScore                 int     `json:"safeScore"`
 }
 type MemberListResp struct {
 	RecordArr  []*UnionMember `json:"recordArr"`
